requirements.txt
streamlit==1.30.0
pandas==2.2.3
streamlit-aggrid
langchain==0.2.17
openai==1.59.8
faiss-cpu==1.7.4
tiktoken==0.5.2
pinecone-client==3.2.2
pinecone-datasets==0.7.0
python-dotenv==1.0.1
pyathena==3.12.2
boto3==1.36.1
PyPDF2==3.0.1
streamlit-option-menu==0.3.12
streamlit-chat==0.1.1
langchain-community==0.2.19
langchain-core==0.2.43